src/requirements.txt
# =============================================================================
# `requirements.txt` installs all dependant packages
# =============================================================================
pip
wheel
setuptools
virtualenv
asyncio
cryptography
asyncua
dataclasses
texttable
docutils
pygments
wifi
RPIO
prettyprint
pyserial
pypubsub
h5pyd
requests
kwargs
adafruit-circuitpython-busdevice
adafruit-circuitpython-am2320
azure.iot.device
azure.identity
azure.keyvault.certificates
azure.keyvault.secrets
azure.keyvault.keys
